+++ src/server/graphql/blockchain/fetchers/transactions.ts
@@ -12,11 +12,14 @@ import {
     getNodeSelectionSetForConnection,
     isDefined,
     prepareChainOrderFilter,
+    prepareNonChainOrderPaginationFilter,
     processPaginatedQueryResult,
     processPaginationArgs,
 } from "../helpers"
 import {
     BlockchainAccountQueryTransactionsArgs,
+    BlockchainAccountQueryTransactions_By_LtArgs,
+    BlockchainQueryTransactions_By_In_MsgArgs,
     BlockchainQueryTransactionsArgs,
     BlockchainTransaction,
     BlockchainTransactionsConnection,
@@ -59,6 +62,46 @@ export async function resolve_transaction(
     )) as BlockchainTransaction[]
 
     return queryResult[0]
+}
+
+export async function resolve_transactions_by_in_msg(
+    args: BlockchainQueryTransactions_By_In_MsgArgs,
+    context: QRequestContext,
+    info: GraphQLResolveInfo,
+    traceSpan: QTraceSpan,
+) {
+    const maxJoinDepth = 2
+
+    const selectionSet = info.fieldNodes[0].selectionSet
+    const returnExpression = config.transactions.buildReturnExpression(
+        selectionSet,
+        context,
+        maxJoinDepth,
+        "doc",
+    )
+
+    // query
+    const params = new QParams({
+        stringifyKeyInAqlComparison:
+            context.services.config.queries.filter.stringifyKeyInAqlComparison,
+    })
+    const query =
+        "FOR doc IN transactions " +
+        `FILTER doc.in_msg == @${params.add(args.msg_hash)} ` +
+        "SORT doc.lt ASC LIMIT 50 " +
+        `RETURN ${returnExpression}`
+    const queryResult = (await context.services.data.query(
+        required(context.services.data.transactions.provider),
+        {
+            text: query,
+            vars: params.values,
+            orderBy: [],
+            request: context,
+            traceSpan,
+        },
+    )) as BlockchainTransaction[]
+
+    return queryResult
 }
 
 export async function resolve_blockchain_transactions(
@@ -205,12 +248,88 @@ export async function resolve_account_transactions(
             archive: args.archive ?? undefined,
         },
     )) as BlockchainTransaction[]
-
     return (await processPaginatedQueryResult(
         queryResult,
         limit,
         direction,
         "chain_order",
+        async r => {
+            await config.transactions.fetchJoins(
+                r,
+                selectionSet,
+                context,
+                traceSpan,
+                maxJoinDepth,
+            )
+        },
+    )) as BlockchainTransactionsConnection
+}
+
+export async function resolve_account_transactions_by_lt(
+    account_address: string,
+    args: BlockchainAccountQueryTransactions_By_LtArgs,
+    context: QRequestContext,
+    info: GraphQLResolveInfo,
+    traceSpan: QTraceSpan,
+) {
+    const maxJoinDepth = 2
+    // filters
+    const filters: string[] = []
+    const params = new QParams({
+        stringifyKeyInAqlComparison:
+            context.services.config.queries.filter.stringifyKeyInAqlComparison,
+    })
+
+    await prepareNonChainOrderPaginationFilter(
+        args,
+        params,
+        filters,
+        context,
+        "lt",
+    )
+    filters.push(`doc.account_addr == @${params.add(account_address)}`)
+
+    const { direction, limit } = processPaginationArgs(args)
+
+    const selectionSet = getNodeSelectionSetForConnection(info)
+    const returnExpression = config.transactions.buildReturnExpression(
+        selectionSet,
+        context,
+        maxJoinDepth,
+        "doc",
+        ["lt"],
+    )
+
+    // query
+    const query = `
+        FOR doc IN transactions
+        FILTER ${filters.join(" AND ")}
+        SORT doc.lt ${direction == Direction.Backward ? "DESC" : "ASC"}
+        LIMIT ${limit}
+        RETURN ${returnExpression}
+    `
+    const queryResult = (await context.services.data.query(
+        required(context.services.data.transactions.provider),
+        {
+            text: query,
+            vars: params.values,
+            orderBy: [
+                {
+                    path: "lt",
+                    direction: "ASC",
+                },
+            ],
+            request: context,
+            traceSpan,
+            archive: args.archive ?? undefined,
+        },
+    )) as BlockchainTransaction[]
+
+    return (await processPaginatedQueryResult(
+        queryResult,
+        limit,
+        direction,
+        "lt",
         async r => {
             await config.transactions.fetchJoins(
                 r,